requirements-dev.txt
bddrest
pytest-cov
pytest-pudb
ipython
